Move MangaCoverMetadata to coil
